fix(triggers): address review + CI findings on schema sync
- Asana: legacy project_id now migrates to project_gid via a
  model_validator, so stored project-scoped workflows register against
  ASANA_TASK_CREATED instead of failing; catalog schema key aligned
  to project_gid
- Notion payloads: shared authors description extracted to a module
  constant (SonarCloud duplicate-literal)
- Frontend: retired notion_all_page_events slug removed from both
  trigger registries so it can no longer be selected
- Tests: payload tests split into per-module mirror files under
  tests/unit/models/composio_schemas/ (mutation gate requires them);
  new package export-surface contract test; legacy-migration tests
