+++ apps/web/src/features/workflows/triggers/handlers/notion.tsx
@@ -200,7 +200,6 @@ export const notionTriggerHandler: RegisteredHandler = {
   triggerSlugs: [
     "notion_new_page_in_db",
     "notion_page_updated",
-    "notion_all_page_events",
     "notion_page_content_updated",
   ],
 
@@ -234,7 +233,6 @@ export const notionTriggerHandler: RegisteredHandler = {
     let label = "on notion event";
     if (triggerName === "notion_new_page_in_db") label = "on new page in db";
     if (triggerName === "notion_page_updated") label = "on page updated";
-    if (triggerName === "notion_all_page_events") label = "on any page event";
     if (triggerName === "notion_page_content_updated")
       label = "on page content updated";
